fix-Process-Create-options-Unused
```
    static ProcessBuilder create(Vertx vertx, String command, ProcessOptions options) {
        return create(vertx, command, Collections.emptyList(), new ProcessOptions());
    }
```

I've identified a bug in the create method of the ProcessBuilder class. The method signature suggests that it should use the ProcessOptions options parameter provided by the caller. However, in the current implementation, the method disregards this parameter and instead creates a new ProcessOptions object. This behavior could lead to unexpected issues, as any custom configurations passed through the options parameter are not being utilized.